refactor LoginCardWrapper to AuthCardWrapper changed all relevant login naming to auth
